+++ apps/indexer/config/test/nethermind.exs
@@ -4,5 +4,5 @@ config :indexer,
   json_rpc_named_arguments: [
     transport: EthereumJSONRPC.Mox,
     transport_options: [],
-    variant: EthereumJSONRPC.Parity
+    variant: EthereumJSONRPC.Nethermind
   ]